Implement the GC sharing pass for 32-in-64 using the bitmap to distinguish forwarding pointers and sharing chain. Includes some debugging code that can be removed.  The sharing pass is triggered on every full GC for testing.

+++ libpolyml/winbasicio.cpp
@@ -1362,7 +1362,7 @@ static Handle IO_dispatch_c(TaskData *taskData, Handle args, Handle strm, Handle
         // the low order part of the object address.
         WinStream *stream = *(WinStream **)(strm->WordP());
         if (stream == 0) raise_syscall(taskData, "Stream is closed", STREAMCLOSED);
-        return Make_fixed_precision(taskData, (POLYUNSIGNED)(stream) & 0xfffffff);
+        return Make_fixed_precision(taskData, (POLYUNSIGNED)((uintptr_t)(stream)) & 0xfffffff);
     }
 
     default:
